+++ DuckovCustomModel/MonoBehaviours/DontHideAsEquipment.cs
@@ -1,7 +1,9 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace DuckovCustomModel.MonoBehaviours
 {
+    [Obsolete("This MonoBehaviour is deprecated and no longer serves a purpose.", true)]
     public class DontHideAsEquipment : MonoBehaviour
     {
     }